requirements.txt

Flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.0.1
google-cloud-firestore==2.11.0
